backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
replicate==0.23.1
pydantic==2.6.0
python-dotenv==1.0.1
httpx==0.26.0
